:recycle:[refactor]: 매칭 로직이 담겨있는 GaleShapley 클래스를 algorithm 패키지로 분리
